+++ datanode/sqlstore/ledger_test.go
@@ -19,6 +19,8 @@ import (
 
 	"code.vegaprotocol.io/vega/datanode/entities"
 	"code.vegaprotocol.io/vega/datanode/sqlstore"
+	"code.vegaprotocol.io/vega/datanode/sqlstore/helpers"
+	"code.vegaprotocol.io/vega/protos/vega"
 	"github.com/shopspring/decimal"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -28,15 +30,17 @@ func addTestLedgerEntry(t *testing.T, ledger *sqlstore.Ledger,
 	accountFrom entities.Account,
 	accountTo entities.Account,
 	block entities.Block,
+	quantity int64,
+	transferType entities.LedgerMovementType,
 ) entities.LedgerEntry {
 	t.Helper()
 	ledgerEntry := entities.LedgerEntry{
 		AccountFromID: accountFrom.ID,
 		AccountToID:   accountTo.ID,
-		Quantity:      decimal.NewFromInt(100),
+		Quantity:      decimal.NewFromInt(quantity),
 		VegaTime:      block.VegaTime,
 		TransferTime:  block.VegaTime.Add(-time.Second),
-		Type:          entities.LedgerMovementTypeBondSlashing,
+		Type:          transferType,
 	}
 
 	err := ledger.Add(ledgerEntry)
@@ -48,46 +52,461 @@ func TestLedger(t *testing.T) {
 	defer DeleteEverything()
 	ctx := context.Background()
 
+	// Prepare environment entities.
 	blockStore := sqlstore.NewBlocks(connectionSource)
 	assetStore := sqlstore.NewAssets(connectionSource)
 	accountStore := sqlstore.NewAccounts(connectionSource)
 	partyStore := sqlstore.NewParties(connectionSource)
+	marketStore := sqlstore.NewMarkets(connectionSource)
 	ledgerStore := sqlstore.NewLedger(connectionSource)
 
-	// Account store should be empty to begin with
-	ledgerEntries, err := ledgerStore.GetAll()
+	// Setup 4 assets
+	asset1 := addTestAsset(t, assetStore, addTestBlock(t, blockStore))
+	asset2 := addTestAsset(t, assetStore, addTestBlock(t, blockStore))
+	asset3 := addTestAsset(t, assetStore, addTestBlock(t, blockStore))
+
+	var blocks []entities.Block
+	var parties []entities.Party
+	var markets []entities.Market
+	var accounts []entities.Account
+
+	/*
+		--- env ---
+		block 0		block 1		block 2		block 3		block 4		block 5		block 6		block 7		block 8		block 9		block 10
+		party 0		party 1		party 2		party 3		party 4		party 5		party 6		party 7		party 8		party 9		party 10
+
+		market 0	market 1 	market 2	market 3	market 4	market 5	market 6	market 7	market 8	market 9	market 10
+
+		--- accounts ---
+		accounts[0] => asset1, parties[0], markets[0], vega.AccountType_ACCOUNT_TYPE_GENERAL
+		accounts[1] => asset1, parties[0], markets[1], vega.AccountType_ACCOUNT_TYPE_GENERAL
+		accounts[2] => asset1, parties[1], markets[1], vega.AccountType_ACCOUNT_TYPE_GENERAL
+		accounts[3] => asset1, parties[1], markets[2], vega.AccountType_ACCOUNT_TYPE_GENERAL
+
+		accounts[4] => asset2, parties[2], markets[2], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+		accounts[5] => asset2, parties[2], markets[3], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+		accounts[6] => asset2, parties[3], markets[3], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+		accounts[7] => asset2, parties[3], markets[4], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+		accounts[8] => asset2, parties[4], markets[4], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+		accounts[9] => asset2, parties[4], markets[5], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+		accounts[10] => asset2, parties[5], markets[5], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+		accounts[11] => asset2, parties[5], markets[6], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+
+		accounts[12] => asset3, parties[6], markets[6], vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY
+		accounts[13] => asset3, parties[6], markets[7], vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY
+		accounts[14] => asset3, parties[7], markets[7], vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY
+		accounts[15] => asset3, parties[7], markets[8], vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY
+		accounts[16] => asset3, parties[8], markets[8], vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY
+		accounts[17] => asset3, parties[8], markets[9], vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY
+		accounts[18] => asset3, parties[9], markets[9], vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY
+		accounts[19] => asset3, parties[9], markets[10], vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY
+		accounts[20] => asset3, parties[10], markets[10], vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY
+		accounts[21] => asset3, parties[10], markets[11], vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY
+	*/
+	for i := 0; i < 17; i++ {
+		blocks = append(blocks, addTestBlockForTime(t, blockStore, time.Now().Add((-26*time.Hour)-(time.Duration(5-i)*time.Second))))
+		parties = append(parties, addTestParty(t, partyStore, blocks[i]))
+		markets = append(markets, helpers.GenerateMarkets(t, 1, blocks[0], marketStore)[0])
+	}
+
+	for i := 0; i < 11; i++ {
+		var mt int
+		if i < 11-1 {
+			mt = i + 1
+		} else {
+			mt = i - 1
+		}
+
+		if i < 2 {
+			// accounts 0, 1, 2, 3
+			accountFrom := helpers.AddTestAccountWithMarketAndType(t, accountStore, parties[i], asset1, blocks[i], markets[i].ID, vega.AccountType_ACCOUNT_TYPE_GENERAL)
+			accountTo := helpers.AddTestAccountWithMarketAndType(t, accountStore, parties[i], asset1, blocks[i], markets[mt].ID, vega.AccountType_ACCOUNT_TYPE_GENERAL)
+			accounts = append(accounts, accountFrom)
+			accounts = append(accounts, accountTo)
+			continue
+		}
+
+		// accounts 4, 5, 6, 7, 8, 9, 10, 11
+		if i >= 2 && i < 6 {
+			accountFrom := helpers.AddTestAccountWithMarketAndType(t, accountStore, parties[i], asset2, blocks[i], markets[i].ID, vega.AccountType_ACCOUNT_TYPE_INSURANCE)
+			accountTo := helpers.AddTestAccountWithMarketAndType(t, accountStore, parties[i], asset2, blocks[i], markets[mt].ID, vega.AccountType_ACCOUNT_TYPE_INSURANCE)
+			accounts = append(accounts, accountFrom)
+			accounts = append(accounts, accountTo)
+			continue
+		}
+
+		// accounts 12, 13, 14, 15, 16, 17, 18, 19, 20, 21
+		accountFrom := helpers.AddTestAccountWithMarketAndType(t, accountStore, parties[i], asset3, blocks[i], markets[i].ID, vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY)
+		accountTo := helpers.AddTestAccountWithMarketAndType(t, accountStore, parties[i], asset3, blocks[i], markets[mt].ID, vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY)
+		accounts = append(accounts, accountFrom)
+		accounts = append(accounts, accountTo)
+	}
+
+	/*
+		--- Transfers ---
+		Asset1:
+		accounts[0]->accounts[1] => asset1, parties[0], markets[0-1], vega.AccountType_ACCOUNT_TYPE_GENERAL
+		accounts[2]->accounts[3] => asset1, parties[1], markets[1-2], vega.AccountType_ACCOUNT_TYPE_GENERAL
+
+		Asset2:
+		accounts[4]->accounts[5] => asset2, parties[2], markets[2-3], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+		accounts[6]->accounts[7] => asset2, parties[3], markets[3-4], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+		accounts[6]->accounts[7] => asset2, parties[3], markets[3-4], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+		accounts[6]->accounts[7] => asset2, parties[3], markets[3-4], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+		accounts[8]->accounts[9] => asset2, parties[4], markets[4-5], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+		accounts[10]->accounts[11] => asset2, parties[5], markets[5-6], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+
+		accounts[5]->accounts[10] => asset2, parties[2-5], markets[3-5], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+
+		accounts[5]->accounts[11] => asset2, parties[2-5], markets[3-6], vega.AccountType_ACCOUNT_TYPE_INSURANCE
+
+		Asset3:
+		accounts[14]->accounts[16] => asset3, parties[7-8], markets[7-8], vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY
+
+		accounts[17]->accounts[15] => asset3, parties[8-7], markets[9-8], vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY
+
+		accounts[21]->accounts[15] => asset3, parties[10-7], markets[11-8], vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY
+	*/
+	var ledgerEntries []entities.LedgerEntry
+	ledgerEntries = append(ledgerEntries, addTestLedgerEntry(t, ledgerStore, accounts[0], accounts[1], blocks[1], int64(15), entities.LedgerMovementTypeBondSlashing))
+	ledgerEntries = append(ledgerEntries, addTestLedgerEntry(t, ledgerStore, accounts[2], accounts[3], blocks[2], int64(10), entities.LedgerMovementTypeBondSlashing))
+	ledgerEntries = append(ledgerEntries, addTestLedgerEntry(t, ledgerStore, accounts[4], accounts[5], blocks[3], int64(25), entities.LedgerMovementTypeBondSlashing))
+	ledgerEntries = append(ledgerEntries, addTestLedgerEntry(t, ledgerStore, accounts[6], accounts[7], blocks[4], int64(80), entities.LedgerMovementTypeBondSlashing))
+	ledgerEntries = append(ledgerEntries, addTestLedgerEntry(t, ledgerStore, accounts[8], accounts[9], blocks[5], int64(1), entities.LedgerMovementTypeDeposit))
+	ledgerEntries = append(ledgerEntries, addTestLedgerEntry(t, ledgerStore, accounts[10], accounts[11], blocks[6], int64(40), entities.LedgerMovementTypeDeposit))
+	ledgerEntries = append(ledgerEntries, addTestLedgerEntry(t, ledgerStore, accounts[14], accounts[16], blocks[7], int64(12), entities.LedgerMovementTypeDeposit))
+	ledgerEntries = append(ledgerEntries, addTestLedgerEntry(t, ledgerStore, accounts[17], accounts[15], blocks[8], int64(14), entities.LedgerMovementTypeDeposit))
+	ledgerEntries = append(ledgerEntries, addTestLedgerEntry(t, ledgerStore, accounts[21], accounts[15], blocks[9], int64(28), entities.LedgerMovementTypeDeposit))
+	ledgerEntries = append(ledgerEntries, addTestLedgerEntry(t, ledgerStore, accounts[5], accounts[11], blocks[10], int64(3), entities.LedgerMovementTypeRewardPayout))
+	ledgerEntries = append(ledgerEntries, addTestLedgerEntry(t, ledgerStore, accounts[5], accounts[10], blocks[11], int64(5), entities.LedgerMovementTypeRewardPayout))
+	ledgerEntries = append(ledgerEntries, addTestLedgerEntry(t, ledgerStore, accounts[6], accounts[7], blocks[12], int64(9), entities.LedgerMovementTypeRewardPayout))
+	_ = append(ledgerEntries, addTestLedgerEntry(t, ledgerStore, accounts[6], accounts[7], blocks[13], int64(41), entities.LedgerMovementTypeRewardPayout))
+
+	tStart, _ := time.Parse("2006 Jan 02 15:04:05", "2012 Dec 07 00:00:00")
+	tEnd := time.Now()
+
+	t.Run("get all ledger records", func(t *testing.T) {
+		// Account store should be empty to begin with
+		ledgerEntries, err := ledgerStore.GetAll()
+		assert.NoError(t, err)
+		assert.Empty(t, ledgerEntries)
+	})
+
+	_, err := ledgerStore.Flush(ctx)
 	assert.NoError(t, err)
-	assert.Empty(t, ledgerEntries)
 
-	block := addTestBlock(t, blockStore)
-	asset := addTestAsset(t, assetStore, block)
-	party := addTestParty(t, partyStore, block)
-	accountFrom := addTestAccount(t, accountStore, party, asset, block)
-	accountTo := addTestAccount(t, accountStore, party, asset, block)
-	firstLedgerEntry := addTestLedgerEntry(t, ledgerStore, accountFrom, accountTo, block)
+	t.Run("ledger entries with no filters", func(t *testing.T) {
+		// Set filters for AccountFrom and AcountTo IDs
+		filter := &entities.LedgerEntryFilter{
+			AccountFromFilters: []entities.AccountFilter{},
+			AccountToFilters:   []entities.AccountFilter{},
+		}
 
-	_, err = ledgerStore.Flush(ctx)
-	assert.NoError(t, err)
+		entries, _, err := ledgerStore.Query(
+			filter,
+			&sqlstore.GroupOptions{},
+			entities.DateRange{Start: &tStart, End: &tEnd},
+			entities.CursorPagination{},
+		)
 
-	// Add it again twice; we're allowed multiple ledger entries with the same parameters in the same block as well
-	// as acroos block
-	block2 := addTestBlock(t, blockStore)
-	addTestLedgerEntry(t, ledgerStore, accountFrom, accountTo, block2)
-	addTestLedgerEntry(t, ledgerStore, accountFrom, accountTo, block2)
+		assert.NoError(t, err)
+		// Output entries for accounts positions:
+		// *
+		assert.Equal(t, 13, len(*entries))
+	})
 
-	_, err = ledgerStore.Flush(ctx)
-	assert.NoError(t, err)
+	t.Run("query ledger entries with filters", func(t *testing.T) {
+		t.Run("by accountFrom filter", func(t *testing.T) {
+			// Set filters for AccountFrom and AcountTo IDs
+			filter := &entities.LedgerEntryFilter{
+				AccountFromFilters: []entities.AccountFilter{
+					{AssetID: asset1.ID},
+					{
+						AssetID:  asset2.ID,
+						PartyIDs: []entities.PartyID{parties[3].ID},
+					},
+				},
+				AccountToFilters: []entities.AccountFilter{},
+			}
 
-	// Query and check we've got back an asset the same as the one we put in, once we give it an ID
-	ledgerEntryTime := entities.CreateLedgerEntryTime(block.VegaTime, 0)
+			entries, _, err := ledgerStore.Query(
+				filter,
+				&sqlstore.GroupOptions{},
+				entities.DateRange{Start: &tStart, End: &tEnd},
+				entities.CursorPagination{},
+			)
 
-	fetchedLedgerEntry, err := ledgerStore.GetByLedgerEntryTime(ledgerEntryTime)
-	assert.NoError(t, err)
-	firstLedgerEntry.LedgerEntryTime = ledgerEntryTime
-	assert.Equal(t, firstLedgerEntry, fetchedLedgerEntry)
+			assert.NoError(t, err)
+			// Output entries for accounts positions:
+			// 0->1, 2->3
+			// 6->7, 6->7, 6->7
+			assert.Equal(t, 5, len(*entries))
 
-	// We should have added three entries in total
-	ledgerEntriesAfter, err := ledgerStore.GetAll()
-	assert.NoError(t, err)
-	assert.Len(t, ledgerEntriesAfter, 3)
+			filter.AccountFromFilters[1].PartyIDs = append(filter.AccountFromFilters[1].PartyIDs, parties[4].ID)
+
+			entries, _, err = ledgerStore.Query(
+				filter,
+				&sqlstore.GroupOptions{},
+				entities.DateRange{Start: &tStart, End: &tEnd},
+				entities.CursorPagination{},
+			)
+
+			assert.NoError(t, err)
+			// Output entries for accounts positions:
+			// 0->1, 2->3
+			// 6->7, 6->7, 6->7, 8->9
+			assert.Equal(t, 6, len(*entries))
+
+			filter.AccountFromFilters[1].AccountTypes = []vega.AccountType{vega.AccountType_ACCOUNT_TYPE_GENERAL}
+
+			entries, _, err = ledgerStore.Query(
+				filter,
+				&sqlstore.GroupOptions{},
+				entities.DateRange{Start: &tStart, End: &tEnd},
+				entities.CursorPagination{},
+			)
+
+			assert.NoError(t, err)
+			// Output entries for accounts positions:
+			// 0->1, 2->3
+			assert.Equal(t, 2, len(*entries))
+		})
+
+		t.Run("by accountTo filter", func(t *testing.T) {
+			// Set filters for AccountFrom and AcountTo IDs
+			filter := &entities.LedgerEntryFilter{
+				AccountFromFilters: []entities.AccountFilter{},
+				AccountToFilters: []entities.AccountFilter{
+					{AssetID: asset1.ID},
+					{
+						AssetID:  asset2.ID,
+						PartyIDs: []entities.PartyID{parties[3].ID},
+					},
+				},
+			}
+
+			entries, _, err := ledgerStore.Query(
+				filter,
+				&sqlstore.GroupOptions{},
+				entities.DateRange{Start: &tStart, End: &tEnd},
+				entities.CursorPagination{},
+			)
+
+			assert.NoError(t, err)
+			// Output entries for accounts positions:
+			// 0->1, 2->3
+			// 6->7, 6->7, 6->7
+			assert.Equal(t, 5, len(*entries))
+
+			filter.AccountToFilters[1].PartyIDs = append(filter.AccountToFilters[1].PartyIDs, parties[4].ID)
+
+			entries, _, err = ledgerStore.Query(
+				filter,
+				&sqlstore.GroupOptions{},
+				entities.DateRange{Start: &tStart, End: &tEnd},
+				entities.CursorPagination{},
+			)
+
+			assert.NoError(t, err)
+			// Output entries for accounts positions:
+			// 0->1, 2->3
+			// 6->7, 6->7, 6->7, 8->9
+			assert.Equal(t, 6, len(*entries))
+
+			filter.AccountToFilters[1].AccountTypes = []vega.AccountType{vega.AccountType_ACCOUNT_TYPE_GENERAL, vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY}
+
+			entries, _, err = ledgerStore.Query(
+				filter,
+				&sqlstore.GroupOptions{},
+				entities.DateRange{Start: &tStart, End: &tEnd},
+				entities.CursorPagination{},
+			)
+
+			assert.NoError(t, err)
+			// Output entries for accounts positions:
+			// 0->1, 2->3
+			assert.Equal(t, 2, len(*entries))
+		})
+
+		t.Run("by accountFrom+accountTo filters", func(t *testing.T) {
+			t.Run("open", func(t *testing.T) {
+				// Set filters for AccountFrom and AcountTo IDs
+				filter := &entities.LedgerEntryFilter{
+					AccountFromFilters: []entities.AccountFilter{
+						{AssetID: asset1.ID},
+						{AssetID: asset2.ID},
+					},
+					AccountToFilters: []entities.AccountFilter{
+						{AssetID: asset2.ID},
+						{AssetID: asset3.ID},
+					},
+				}
+
+				entries, _, err := ledgerStore.Query(
+					filter,
+					&sqlstore.GroupOptions{},
+					entities.DateRange{Start: &tStart, End: &tEnd},
+					entities.CursorPagination{},
+				)
+
+				assert.NoError(t, err)
+				// Output entries for accounts positions:
+				// 0->1, 2->3
+				// 4->5, 6->7, 6->7, 6->7, 8->9, 5->10, 5->11, 10->11
+				// 14->16, 17->15, 21->15
+				assert.Equal(t, 13, len(*entries))
+
+				filter.AccountToFilters[1].PartyIDs = append(filter.AccountToFilters[1].PartyIDs, []entities.PartyID{parties[3].ID, parties[4].ID}...)
+				entries, _, err = ledgerStore.Query(
+					filter,
+					&sqlstore.GroupOptions{},
+					entities.DateRange{Start: &tStart, End: &tEnd},
+					entities.CursorPagination{},
+				)
+
+				assert.NoError(t, err)
+				// Output entries for accounts positions:
+				// 0->1, 2->3
+				// 4->5, 6->7, 6->7, 6->7, 8->9
+				// 14->16, 17->15, 21->15
+				assert.Equal(t, 10, len(*entries))
+
+				filter.AccountToFilters[1].AccountTypes = []vega.AccountType{vega.AccountType_ACCOUNT_TYPE_GENERAL, vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY}
+
+				entries, _, err = ledgerStore.Query(
+					filter,
+					&sqlstore.GroupOptions{},
+					entities.DateRange{Start: &tStart, End: &tEnd},
+					entities.CursorPagination{},
+				)
+
+				assert.NoError(t, err)
+				// Output entries for accounts positions:
+				// 0->1, 2->3
+				// 4->5, 6->7, 6->7, 6->7, 8->9
+				// 14->16, 17->15, 21->15
+				assert.Equal(t, 10, len(*entries))
+			})
+
+			t.Run("closed", func(t *testing.T) {
+				// Set filters for AccountFrom and AcountTo IDs
+				filter := &entities.LedgerEntryFilter{
+					AccountFromFilters: []entities.AccountFilter{
+						{AssetID: asset1.ID},
+						{AssetID: asset2.ID},
+					},
+					AccountToFilters: []entities.AccountFilter{
+						{AssetID: asset2.ID},
+						{AssetID: asset3.ID},
+					},
+				}
+
+				filter.CloseOnAccountFilters = true
+				entries, _, err := ledgerStore.Query(
+					filter,
+					&sqlstore.GroupOptions{},
+					entities.DateRange{Start: &tStart, End: &tEnd},
+					entities.CursorPagination{},
+				)
+
+				assert.NoError(t, err)
+				// Output entries for accounts positions -> should output transfers for asset2 only:
+				// 4->5, 6->7, 6->7, 6->7, 8->9, 5->10, 5->11, 10->11
+				assert.Equal(t, 8, len(*entries))
+
+				// Add some grouping options
+				filter.AccountToFilters = append(filter.AccountToFilters, []entities.AccountFilter{{AssetID: asset3.ID}}...)
+				entries, _, err = ledgerStore.Query(
+					filter,
+					&sqlstore.GroupOptions{
+						ByAccountField: []entities.AccountField{entities.AccountFieldType},
+					},
+					entities.DateRange{Start: &tStart, End: &tEnd},
+					entities.CursorPagination{},
+				)
+
+				assert.NoError(t, err)
+				// Output entries for accounts positions:
+				// 4->5, 6->7, 6->7, 6->7, 8->9, 10->11, 5->10, 5->11 - ACCOUNT_TYPE_INSURANCE
+				assert.Equal(t, 8, len(*entries))
+
+				filter.AccountToFilters[1].AccountTypes = []vega.AccountType{vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY}
+
+				entries, _, err = ledgerStore.Query(
+					filter,
+					&sqlstore.GroupOptions{},
+					entities.DateRange{Start: &tStart, End: &tEnd},
+					entities.CursorPagination{},
+				)
+
+				assert.NoError(t, err)
+				// Output entries for accounts positions:
+				// 4->5, 6->7, 6->7, 6->7, 8->9, 10->11, 5->10, 5->11
+				assert.Equal(t, 8, len(*entries))
+			})
+		})
+
+		t.Run("by account filters+transferType", func(t *testing.T) {
+			// open on account filters
+			// Set filters for AccountFrom and AcountTo IDs
+			filter := &entities.LedgerEntryFilter{
+				AccountFromFilters: []entities.AccountFilter{
+					{AssetID: asset1.ID},
+					{AssetID: asset2.ID},
+				},
+				AccountToFilters: []entities.AccountFilter{
+					{AssetID: asset2.ID},
+					{AssetID: asset3.ID},
+				},
+				TransferTypes: []entities.LedgerMovementType{
+					entities.LedgerMovementTypeDeposit,
+				},
+			}
+
+			entries, _, err := ledgerStore.Query(
+				filter,
+				&sqlstore.GroupOptions{},
+				entities.DateRange{Start: &tStart, End: &tEnd},
+				entities.CursorPagination{},
+			)
+
+			assert.NoError(t, err)
+			// Output entries for accounts positions -> should output transfers for asset2 only:
+			// 8->9, 10->11, 14->16, 17->15, 21->15
+			assert.Equal(t, 5, len(*entries))
+
+			// closed on account filters
+			filter.CloseOnAccountFilters = true
+			entries, _, err = ledgerStore.Query(
+				filter,
+				&sqlstore.GroupOptions{},
+				entities.DateRange{Start: &tStart, End: &tEnd},
+				entities.CursorPagination{},
+			)
+
+			assert.NoError(t, err)
+			// Output entries for accounts positions -> should output transfers for asset2 only:
+			// 8->9, 10->11
+			assert.Equal(t, 2, len(*entries))
+
+			filter.AccountToFilters = []entities.AccountFilter{
+				{
+					AssetID:      asset3.ID,
+					AccountTypes: []vega.AccountType{vega.AccountType_ACCOUNT_TYPE_FEES_LIQUIDITY},
+				},
+			}
+
+			entries, _, err = ledgerStore.Query(
+				filter,
+				&sqlstore.GroupOptions{},
+				entities.DateRange{Start: &tStart, End: &tEnd},
+				entities.CursorPagination{},
+			)
+
+			assert.NoError(t, err)
+			// Output entries for accounts positions -> should output transfers for asset2 only:
+			// 0
+			assert.Equal(t, 0, len(*entries))
+		})
+	})
 }